requirements.txt
uvicorn[standard]==0.15.0
fastapi==0.95.1
python-decouple==3.8
aioredis==2.0.1
Inject==5.0.0
gunicorn==20.1.0
ujson==5.8.0
SQLAlchemy==2.0.16
psycopg2-binary==2.9.6
pyhumps==3.8.0
splitio_client==9.4.2
boto3==1.26.153
pydantic==1.10.9
httpx==0.24.1


# lib to study in future with quantum programming
qiskit
